% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{database_key}
\alias{database_key}
\title{View Available Geneset Databases}
\usage{
database_key(database_name = "")
}
\arguments{
\item{database_name}{The input for this paramter needs to be "all", or the description of the database of interest.}
}
\value{
If the input for database_name is "all", the function will return the descriptons for all databases. If the user inputs the descripton of a database, the function will return the file name of the database of interest.
}
\description{
The function database_key returns the name of the rda file the user may be interested in given the common name of the geneset database.
The function may also return the name of all databases if the input is "all".
}
\examples{
database_key("all")

database_key("hallmark all")

database_key("motif transcription factor targets")
}
