% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{get_genesets}
\alias{get_genesets}
\title{View Gene Sets within Database}
\usage{
get_genesets(database_file = "")
}
\arguments{
\item{database_file}{Name of the rda file for the geneset database within the package}
}
\value{
The names of the genesets within the database
}
\description{
This function takes one input, the file name for the database, and returns the names of the sets within the database.
}
