% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{get_genesymbols}
\alias{get_genesymbols}
\title{Get Gene Symbols}
\usage{
get_genesymbols(database_file = "")
}
\arguments{
\item{database_file}{Name of the rda file for the geneset database within the package.}
}
\value{
The gene symbols within each gene set in a database.
}
\description{
The get_genesymbols function returns the gene symbols within each gene set given the file name of the database of interest.
}
\examples{
get_genesymbols(hallmark.gs)

get_genesymbols(C3.tft.gs)

get_genesymbols(custom_db)
}
