% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEA.1.0.sei3.R
\name{GSEA_Source}
\alias{GSEA_Source}
\alias{GSEA.GeneRanking}
\alias{GSEA.EnrichmentScore}
\alias{OLD.GSEA.EnrichmentScore}
\alias{GSEA.EnrichmentScore2}
\alias{GSEA.HeatMapPlot}
\alias{GSEA.Res2Frame}
\alias{GSEA.Gct2Frame}
\alias{GSEA.Gct2Frame2}
\alias{GSEA.ReadClsFile}
\alias{GSEA.Threshold}
\alias{GSEA.VarFilter}
\alias{GSEA.NormalizeRows}
\alias{GSEA.NormalizeCols}
\alias{GSEA}
\title{GSEA Source Code Functions}
\usage{
GSEA.GeneRanking(A, class.labels, gene.labels, nperm,
  permutation.type = 0, sigma.correction = "GeneCluster",
  fraction = 1, replace = F, reverse.sign = F, abs.val = F)

GSEA.EnrichmentScore(gene.list, gene.set, weighted.score.type = 1,
  correl.vector = NULL)

OLD.GSEA.EnrichmentScore(gene.list, gene.set)

GSEA.EnrichmentScore2(gene.list, gene.set, weighted.score.type = 1,
  correl.vector = NULL)

GSEA.HeatMapPlot(V, row.names = F, col.labels, col.classes,
  col.names = F, main = " ", xlab = " ", ylab = " ")

GSEA.Res2Frame(filename = "NULL")

GSEA.Gct2Frame(filename = "NULL")

GSEA.Gct2Frame2(filename = "NULL")

GSEA.ReadClsFile(file = "NULL")

GSEA.Threshold(V, thres, ceil)

GSEA.VarFilter(V, fold, delta, gene.names = "NULL")

GSEA.NormalizeRows(V)

GSEA.NormalizeCols(V)

GSEA(input.ds, input.cls, gene.ann = "", gs.db, gs.ann = "",
  output.directory = "", output.directory2 = "",
  doc.string = "GSEA.analysis", non.interactive.run = F,
  reshuffling.type = "sample.labels", nperm = 1000,
  weighted.score.type = 1, nom.p.val.threshold = -1,
  fwer.p.val.threshold = -1, fdr.q.val.threshold = 0.25, topgs = 10,
  adjust.FDR.q.val = F, gs.size.threshold.min = 25,
  gs.size.threshold.max = 500, reverse.sign = F, preproc.type = 0,
  random.seed = 123456, perm.type = 0, fraction = 1, replace = F,
  save.intermediate.results = F, OLD.GSEA = F,
  use.fast.enrichment.routine = T, abs.val = F)
}
\description{
These functions were written by the Broad Institute and published as a part of their attempt to make their Gene Set Enrichment Analysis able to be run in R. These functions are used within the functions written in this package. They are not meant to be used by the user but are made available if they need to be used.
}
\references{
Subramanian, Tamayo, et al. (2005), PNAS 102, 15545-15550, http://www.broad.mit.edu/gsea/
}
