% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{add_to_database}
\alias{add_to_database}
\title{Add to Existing Gene Set Database}
\usage{
add_to_database(database = "", addition = "")
}
\arguments{
\item{database}{formatted existing database.}

\item{addition}{gene sets to be added to existing database.}
}
\value{
a formatted database with new sets included.
}
\description{
The function add_to_database returns a database comprised of an existing database and new data, which are paramters provded by the user. The dataset that results from this function will be formatted to be compatible with GSEA functions.
}
