% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{add_to_database}
\alias{add_to_database}
\title{Add to Existing Gene set Database}
\usage{
add_to_database(database = "", addition = "")
}
\arguments{
\item{database}{the formatted existing database}

\item{addition}{the gene sets to be added}
}
\value{
a formatted database with new sets included
}
\description{
This function takes two inputs: the formatted existing database and the gene sets to add.It will return the combined database ready to be used in the GSEA function.
}
