% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{create_geneset_db}
\alias{create_geneset_db}
\title{Format New Geneset Daatbase}
\usage{
create_geneset_db(database = "")
}
\arguments{
\item{database}{a list names are the sets within the database and contained within those are genesymbols}
}
\value{
formatted_db a database ready to be saved to the package and used in later GSEA analyses
}
\description{
When given a list with keys as a set name and the data corresponding to a key, the gene symbols corresponding to that key. This function will reformat the data to be saved into the package and to be a working input to the GSEA.plots function.
}
