% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{database_key}
\alias{database_key}
\title{View Available Geneset Databases}
\usage{
database_key(database_name = "")
}
\arguments{
\item{database_name}{Either enter "all" or a string containing the name of one of the existing geneset databases}
}
\value{
Either the names of all existing gene set databases or the file name for the specified database
}
\description{
This function takes one input, the common name of the geneset database you are interested in and returns the name of the rda file that can be used in the GSEA analysis function.
If the input is "all" then all database common names are returned.
}
