% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_input.R
\name{get_genesymbols}
\alias{get_genesymbols}
\title{Get Gene Symbols}
\usage{
get_genesymbols(database_file = "")
}
\arguments{
\item{database_file}{Name of the rda file for the geneset database within the package}
}
\value{
The gene symbols within that gene set
}
\description{
This function takes one input, the file name for the database and and returns the gene symbols within that set.
}
