% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ES.R
\name{plot.ES}
\alias{plot.ES}
\title{GSEA Enrichment Plots}
\usage{
\method{plot}{ES}(list.of.plots = "", plotname = "")
}
\arguments{
\item{list.of.plots}{A list of ggplot items}

\item{plotname}{filename to save the plot as}
}
\description{
Create a pdf of a list of ggplot objects that is obtained from the GSEA analysis. Outputs 4 graphs per page and saves to your working directory
}
